# ENH: Perf improvements to np.sort, np.argsort, np.partition and np.argpartition 