--- a/numpy/_core/src/npysort/simd_qsort.dispatch.cpp
+++ b/numpy/_core/src/npysort/simd_qsort.dispatch.cpp
@@ -11,7 +11,10 @@
 #if defined(NPY_HAVE_AVX512_SKX)
     #include "x86-simd-sort/src/avx512-32bit-qsort.hpp"
     #include "x86-simd-sort/src/avx512-64bit-qsort.hpp"
+/* arg methods are only intended to 64-bit platforms */
+#ifdef NPY_CPU_AMD64
     #include "x86-simd-sort/src/avx512-64bit-argsort.hpp"
+#endif
 #endif
 
 namespace np { namespace qsort_simd {
@@ -19,27 +22,39 @@ namespace np { namespace qsort_simd {
 #if defined(NPY_HAVE_AVX512_SKX)
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(int32_t *arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(uint32_t *arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(int64_t*arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(uint64_t*arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(float *arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(double *arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(QSelect)(int32_t *arr, npy_intp num, npy_intp kth)
 {
@@ -91,27 +106,39 @@ template<> void NPY_CPU_DISPATCH_CURFX(QSort)(double *arr, intptr_t size)
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(int32_t *arr, npy_intp *arg, npy_intp size)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(uint32_t *arr, npy_intp *arg, npy_intp size)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(int64_t *arr, npy_intp *arg, npy_intp size)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(uint64_t *arr, npy_intp *arg, npy_intp size)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(float *arr, npy_intp *arg, npy_intp size)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+#endif
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(double *arr, npy_intp *arg, npy_intp size)
 {
+#ifdef NPY_CPU_AMD64
     avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+#endif
 }
 #endif  // NPY_HAVE_AVX512_SKX