--- a/numpy/_core/src/npysort/simd_qsort.dispatch.cpp
+++ b/numpy/_core/src/npysort/simd_qsort.dispatch.cpp
@@ -19,27 +19,27 @@ namespace np { namespace qsort_simd {
 #if defined(NPY_HAVE_AVX512_SKX)
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(int32_t *arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
-    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+    avx512_argselect(arr, reinterpret_cast<size_t*>(arg), kth, num);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(uint32_t *arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
-    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+    avx512_argselect(arr, reinterpret_cast<size_t*>(arg), kth, num);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(int64_t*arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
-    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+    avx512_argselect(arr, reinterpret_cast<size_t*>(arg), kth, num);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(uint64_t*arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
-    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+    avx512_argselect(arr, reinterpret_cast<size_t*>(arg), kth, num);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(float *arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
-    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+    avx512_argselect(arr, reinterpret_cast<size_t*>(arg), kth, num);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSelect)(double *arr, npy_intp* arg, npy_intp num, npy_intp kth)
 {
-    avx512_argselect(arr, reinterpret_cast<int64_t*>(arg), kth, num);
+    avx512_argselect(arr, reinterpret_cast<size_t*>(arg), kth, num);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(QSelect)(int32_t *arr, npy_intp num, npy_intp kth)
 {
@@ -65,53 +65,53 @@ template<> void NPY_CPU_DISPATCH_CURFX(QSelect)(double *arr, npy_intp num, npy_i
 {
     avx512_qselect(arr, kth, num, true);
 }
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(int32_t *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(int32_t *arr, npy_intp size)
 {
     avx512_qsort(arr, size);
 }
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(uint32_t *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(uint32_t *arr, npy_intp size)
 {
     avx512_qsort(arr, size);
 }
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(int64_t *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(int64_t *arr, npy_intp size)
 {
     avx512_qsort(arr, size);
 }
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(uint64_t *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(uint64_t *arr, npy_intp size)
 {
     avx512_qsort(arr, size);
 }
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(float *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(float *arr, npy_intp size)
 {
     avx512_qsort(arr, size);
 }
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(double *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(double *arr, npy_intp size)
 {
     avx512_qsort(arr, size);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(int32_t *arr, npy_intp *arg, npy_intp size)
 {
-    avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+    avx512_argsort(arr, reinterpret_cast<size_t*>(arg), size);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(uint32_t *arr, npy_intp *arg, npy_intp size)
 {
-    avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+    avx512_argsort(arr, reinterpret_cast<size_t*>(arg), size);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(int64_t *arr, npy_intp *arg, npy_intp size)
 {
-    avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+    avx512_argsort(arr, reinterpret_cast<size_t*>(arg), size);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(uint64_t *arr, npy_intp *arg, npy_intp size)
 {
-    avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+    avx512_argsort(arr, reinterpret_cast<size_t*>(arg), size);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(float *arr, npy_intp *arg, npy_intp size)
 {
-    avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+    avx512_argsort(arr, reinterpret_cast<size_t*>(arg), size);
 }
 template<> void NPY_CPU_DISPATCH_CURFX(ArgQSort)(double *arr, npy_intp *arg, npy_intp size)
 {
-    avx512_argsort(arr, reinterpret_cast<int64_t*>(arg), size);
+    avx512_argsort(arr, reinterpret_cast<size_t*>(arg), size);
 }
 #endif  // NPY_HAVE_AVX512_SKX