--- a/numpy/_core/src/npysort/simd_qsort.hpp
+++ b/numpy/_core/src/npysort/simd_qsort.hpp
@@ -8,7 +8,7 @@ namespace np { namespace qsort_simd {
 #ifndef NPY_DISABLE_OPTIMIZATION
     #include "simd_qsort.dispatch.h"
 #endif
-NPY_CPU_DISPATCH_DECLARE(template <typename T> void QSort, (T *arr, intptr_t size))
+NPY_CPU_DISPATCH_DECLARE(template <typename T> void QSort, (T *arr, npy_intp size))
 NPY_CPU_DISPATCH_DECLARE(template <typename T> void QSelect, (T* arr, npy_intp num, npy_intp kth))
 NPY_CPU_DISPATCH_DECLARE(template <typename T> void ArgQSort, (T *arr, npy_intp* arg, npy_intp size))
 NPY_CPU_DISPATCH_DECLARE(template <typename T> void ArgQSelect, (T *arr, npy_intp* arg, npy_intp kth, npy_intp size))
@@ -16,7 +16,7 @@ NPY_CPU_DISPATCH_DECLARE(template <typename T> void ArgQSelect, (T *arr, npy_int
 #ifndef NPY_DISABLE_OPTIMIZATION
     #include "simd_qsort_16bit.dispatch.h"
 #endif
-NPY_CPU_DISPATCH_DECLARE(template <typename T> void QSort, (T *arr, intptr_t size))
+NPY_CPU_DISPATCH_DECLARE(template <typename T> void QSort, (T *arr, npy_intp size))
 NPY_CPU_DISPATCH_DECLARE(template <typename T> void QSelect, (T* arr, npy_intp num, npy_intp kth))
 
 } } // np::qsort_simd