--- a/numpy/_core/src/npysort/simd_qsort_16bit.dispatch.cpp
+++ b/numpy/_core/src/npysort/simd_qsort_16bit.dispatch.cpp
@@ -14,19 +14,19 @@
  * Wrapper function declarations to avoid multiple definitions of
  * avx512_qsort<uint16_t> and avx512_qsort<int16_t>
  */
-void avx512_qsort_uint16(uint16_t*, intptr_t);
-void avx512_qsort_int16(int16_t*, intptr_t);
+void avx512_qsort_uint16(uint16_t*, npy_intp);
+void avx512_qsort_int16(int16_t*, npy_intp);
 void avx512_qselect_uint16(uint16_t*, npy_intp, npy_intp);
 void avx512_qselect_int16(int16_t*, npy_intp, npy_intp);
 
 #elif defined(NPY_HAVE_AVX512_ICL)
     #include "x86-simd-sort/src/avx512-16bit-qsort.hpp"
 /* Wrapper function defintions here: */
-void avx512_qsort_uint16(uint16_t* arr, intptr_t size)
+void avx512_qsort_uint16(uint16_t* arr, npy_intp size)
 {
     avx512_qsort(arr, size);
 }
-void avx512_qsort_int16(int16_t* arr, intptr_t size)
+void avx512_qsort_int16(int16_t* arr, npy_intp size)
 {
     avx512_qsort(arr, size);
 }
@@ -76,7 +76,7 @@ template<> void NPY_CPU_DISPATCH_CURFX(QSelect)(int16_t *arr, npy_intp num, npy_
 /*
  * QSort dispatch functions:
  */
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(Half *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(Half *arr, npy_intp size)
 {
 #if defined(NPY_HAVE_AVX512_SPR)
     avx512_qsort(reinterpret_cast<_Float16*>(arr), size);
@@ -84,7 +84,7 @@ template<> void NPY_CPU_DISPATCH_CURFX(QSort)(Half *arr, intptr_t size)
     avx512_qsort_fp16(reinterpret_cast<uint16_t*>(arr), size);
 #endif
 }
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(uint16_t *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(uint16_t *arr, npy_intp size)
 {
 #if defined(NPY_HAVE_AVX512_SPR)
     avx512_qsort_uint16(arr, size);
@@ -92,7 +92,7 @@ template<> void NPY_CPU_DISPATCH_CURFX(QSort)(uint16_t *arr, intptr_t size)
     avx512_qsort(arr, size);
 #endif
 }
-template<> void NPY_CPU_DISPATCH_CURFX(QSort)(int16_t *arr, intptr_t size)
+template<> void NPY_CPU_DISPATCH_CURFX(QSort)(int16_t *arr, npy_intp size)
 {
 #if defined(NPY_HAVE_AVX512_SPR)
     avx512_qsort_int16(arr, size);