@r-devulap 2023-10-18
Update: We have changed API to use `size_t` for all array indices and the calls to avx512 routines are updated accordingly. That made it easy for us to vectorize np.argpartition and np.argsort on 32-bit platforms (see PR https://github.com/intel/x86-simd-sort/pull/88).