@seberg 2023-10-18
> We have changed API to use size_t for all array indices

So that is technically be broken or need a guard until we would do gh-24888?  Although I guess that the use of that code implies we are not on a niche platform.