@r-devulap 2023-10-18
> > We have changed API to use size_t for all array indices
> 
> So that is technically be broken or need a guard until we would do [gh-24888](https://github.com/numpy/numpy/pull/24888)? Although I guess that the use of that code implies we are not on a niche platform.

I don't think that's necessary. 

1. For the sort and partition methods `npy_intp` will get cast to `size_t`. Array size passed to sort/partition won't be negative and so this implicit cast isn't a problem.
2. For argsort, we need to reinterpret_cast `npy_intp*` to `size_t*`. Whether `npy_intp` is `Py_intptr_t` or `ssize_t` shouldn't make a difference to this cast. 

let me know if my assumptions are wrong.