@r-devulap 2023-10-18
hmm, the macOS x86_64 failure seems hard to figure out. It's using x86_64-apple-darwin13.4.0-clang++ (clang 15.0.7 "clang version 15.0.7") but I don't see this build fail with clang++-15.0.7 locally. 

Fails with: `error: implicit instantiation of undefined template 'zmm_vector<unsigned long>'`. With long as 8 bytes, unsigned long should just default to `uin64_t` and it should find the template definition like all other cases.