@r-devulap 2023-11-27
Changes in NumPy here are pretty minor in this PR. Perhaps we can close this and merge it into a single PR https://github.com/numpy/numpy/pull/25045. Please reopen if you disagree.